+++ mitra_adapters/src/lib.rs
@@ -4,4 +4,5 @@ mod logger;
 pub mod media;
 pub mod payments;
 pub mod permissions;
+pub mod posts;
 pub mod roles;